refactor: rename canisters to backend/frontend in who_am_i examples
Rename internet_identity_app_backend → backend and
internet_identity_app_frontend → frontend in both Motoko and Rust
who_am_i examples, including src directories, .did files, icp.yaml,
Cargo.toml, Makefile, vite.config.js, actor.js, package.json,
.gitignore, and README. Also removes ICP Ninja artifacts (dfx.json,
BUILD.md) and simplifies vite.config.js by dropping the dfx fallback.
