--- motoko/who_am_i/BUILD.md
@@ -1,26 +1,0 @@
-# Continue building locally
-
-Projects deployed through ICP Ninja are temporary; they will only be live for 30 minutes before they are removed. To continue building locally, follow these steps.
-
-### 1. Install developer tools
-
-Install [Node.js](https://nodejs.org/en/download/) and [icp-cli](https://cli.internetcomputer.org):
-
-```bash
-npm install -g @icp-sdk/icp-cli @icp-sdk/ic-wasm
-```
-
-Then navigate into your project's directory that you downloaded from ICP Ninja.
-
-### 2. Deploy locally
-
-Start the local network and deploy the project:
-
-```bash
-icp network start -d
-icp deploy
-```
-
-## Additional examples
-
-Additional code examples and sample applications can be found in the [DFINITY examples repo](https://github.com/dfinity/examples).
