+++ motoko/who_am_i/package.json
@@ -7,6 +7,6 @@
   },
   "type": "module",
   "workspaces": [
-    "src/internet_identity_app_frontend"
+    "src/frontend"
   ]
 }
