+++ motoko/who_am_i/src/frontend/public/favicon.ico
