+++ rust/who_am_i/.gitignore
@@ -1,1 +1,1 @@
-src/internet_identity_app_frontend/src/bindings/
+src/frontend/src/bindings/
