+++ rust/who_am_i/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["src/internet_identity_app_backend"]
+members = ["src/backend"]
 resolver = "2"
