--- rust/who_am_i/dfx.json
@@ -1,41 +1,0 @@
-{
-  "canisters": {
-    "internet_identity_app_backend": {
-      "candid": "src/internet_identity_app_backend/internet_identity_app_backend.did",
-      "package": "internet_identity_app_backend",
-      "type": "rust"
-    },
-    "internet_identity_app_frontend": {
-      "dependencies": ["internet_identity_app_backend"],
-      "internet_identity_app_frontend": {
-        "entrypoint": "src/internet_identity_app_frontend/index.html"
-      },
-      "source": ["src/internet_identity_app_frontend/dist"],
-      "type": "assets"
-    },
-    "internet_identity": {
-      "candid": "https://github.com/dfinity/internet-identity/releases/latest/download/internet_identity.did",
-      "type": "custom",
-      "specified_id": "rdmx6-jaaaa-aaaaa-aaadq-cai",
-      "remote": {
-        "id": {
-          "ic": "rdmx6-jaaaa-aaaaa-aaadq-cai"
-        }
-      },
-      "wasm": "https://github.com/dfinity/internet-identity/releases/latest/download/internet_identity_production.wasm.gz"
-    },
-    "internet_identity_frontend": {
-      "candid": "https://raw.githubusercontent.com/dfinity/internet-identity/refs/heads/main/src/internet_identity_frontend/internet_identity_frontend.did",
-      "type": "custom",
-      "specified_id": "uqzsh-gqaaa-aaaaq-qaada-cai",
-      "remote": {
-        "id": {
-          "ic": "uqzsh-gqaaa-aaaaq-qaada-cai"
-        }
-      },
-      "wasm": "https://github.com/dfinity/internet-identity/releases/latest/download/internet_identity_frontend.wasm.gz",
-      "init_arg": "(record { fetch_root_key = opt true; dev_csp = opt true; backend_canister_id = principal \"rdmx6-jaaaa-aaaaa-aaadq-cai\"; analytics_config = null; related_origins = opt vec { \"http://uqzsh-gqaaa-aaaaq-qaada-cai.localhost:4943\" }; backend_origin = \"http://rdmx6-jaaaa-aaaaa-aaadq-cai.localhost:4943\"; captcha_config = opt record { max_unsolved_captchas = 50 : nat64; captcha_trigger = variant { Static = variant { CaptchaDisabled } } }})"
-    }
-  },
-  "output_env_file": ".env"
-}
