+++ rust/who_am_i/package.json
@@ -3,7 +3,7 @@
     "node": ">=16.0.0",
     "npm": ">=7.0.0"
   },
-  "name": "internet_identity_app",
+  "name": "who-am-i",
   "scripts": {
     "build": "npm run build --workspaces --if-present",
     "prebuild": "npm run prebuild --workspaces --if-present",
@@ -13,6 +13,6 @@
   },
   "type": "module",
   "workspaces": [
-    "src/internet_identity_app_frontend"
+    "src/frontend"
   ]
 }
